updated delete_files.py

+++ samples/delete_files.py
@@ -11,6 +11,9 @@ final_num = int(sys.argv[2])
 for sub_dir in os.listdir(dir):
     file_names = os.listdir(dir + "/"+sub_dir)
     files_num = len(file_names)
+    target_num = final_num-files_num
+    if target_num <= 0:
+        continue
     selected_files = np.random.choice(file_names, final_num-files_num)
     for file in selected_files:
         os.remove(dir+"/"+sub_dir+"/"+file)
